requirements.txt
Faker
psycopg2
python-dotenv
locust